Move float_parser_t to form hb-number-parser.hh
