+++ src/Makefile.sources
@@ -49,6 +49,7 @@ HB_BASE_sources = \
 	hb-meta.hh \
 	hb-mutex.hh \
 	hb-null.hh \
+	hb-number-parser.hh \
 	hb-object.hh \
 	hb-open-file.hh \
 	hb-open-type.hh \
